requirements.txt
python-telegram-bot==20.3
requests
pandas
numpy
matplotlib
beautifulsoup4
lxml
pytz